requirements.txt
pandas
numpy
matplotlib
yfinance
scipy
ta-lib-bin
pydantic
langgraph
aiohttp
plotly
streamlit
